requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
